src/requirements.txt
streamlit
transformers
pandas
matplotlib
google-generativeai
python-dotenv